Backend/requirements.txt
Django==6.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
python-decouple==3.8
django-filter==24.3
setuptools==82.0.0
